pyproject.toml
[project]
name = "redistest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "fastapi>=0.131.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "yarl>=1.22.0",
]
